[TASK] Fix phpstan checkFunctionArgumentTypes errors in ext:backend Tree
This patch fixes incompatible type usage in function arguments
and is preparatory work for introducing native type hints and
strict mode in all core files.

Releases: master, 10.4
Resolves: #92580
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66159
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>

+++ typo3/sysext/backend/Classes/Tree/SortedTreeNodeCollection.php
@@ -56,7 +56,7 @@ class SortedTreeNodeCollection extends TreeNodeCollection
     protected function binarySearch(TreeNode $node, $start, $end)
     {
         if (!$start && $end - $start >= 2 || $end - $start > 2) {
-            $divider = ceil(($end - $start) / 2);
+            $divider = (int)ceil(($end - $start) / 2);
             if ($this->offsetGet($divider)->equals($node)) {
                 return $divider;
             }
